refactor: move SeriesSetPlans into its own module
