+++ query/src/plan.rs
@@ -1,2 +1,3 @@
 pub mod fieldlist;
+pub mod seriesset;
 pub mod stringset;
